+++ dst/jobs/continuous/abortTasks/run.js
@@ -28,11 +28,13 @@ connectMongo_1.connectMongo().then(() => {
             yield pecorino.service.task.abort(RETRY_INTERVAL_MINUTES)({ task: taskRepo });
         }
         catch (error) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
         count -= 1;
     }), INTERVAL_MILLISECONDS);
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });