+++ dst/jobs/continuous/makeTransactionExpired/run.js
@@ -30,11 +30,13 @@ connectMongo_1.connectMongo().then(() => {
             yield transactionRepo.makeExpired({ expires: new Date() });
         }
         catch (error) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
         count -= 1;
     }), INTERVAL_MILLISECONDS);
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });