+++ dst/jobs/continuous/onConfirmedWithdrawTransaction/run.js
@@ -31,11 +31,13 @@ connectMongo_1.connectMongo().then(() => {
             yield pecorino.service.transaction.withdraw.exportTasks(pecorino.factory.transactionStatusType.Confirmed)({ task: taskRepo, transaction: transactionRepo });
         }
         catch (error) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
         countExecute -= 1;
     }), INTERVAL_MILLISECONDS);
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });