+++ dst/jobs/continuous/reexportTransactionTasks/run.js
@@ -31,11 +31,13 @@ connectMongo_1.connectMongo().then(() => {
             yield transactionRepo.reexportTasks(RETRY_INTERVAL_MINUTES);
         }
         catch (error) {
+            // tslint:disable-next-line:no-console
             console.error(error);
         }
         countRetry -= 1;
     }), INTERVAL_MILLISECONDS);
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });