+++ src/jobs/continuous/abortTasks/run.ts
@@ -2,6 +2,7 @@
  * タスク中止
  */
 import * as pecorino from '@pecorino/domain';
+import * as mongoose from 'mongoose';
 
 import { connectMongo } from '../../../connectMongo';
 
@@ -11,7 +12,7 @@ connectMongo().then(() => {
     const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
     const INTERVAL_MILLISECONDS = 500;
     const RETRY_INTERVAL_MINUTES = 10;
-    const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
+    const taskRepo = new pecorino.repository.Task(mongoose.connection);
 
     setInterval(
         async () => {
@@ -24,6 +25,7 @@ connectMongo().then(() => {
             try {
                 await pecorino.service.task.abort(RETRY_INTERVAL_MINUTES)({ task: taskRepo });
             } catch (error) {
+                // tslint:disable-next-line:no-console
                 console.error(error);
             }
 
@@ -32,6 +34,7 @@ connectMongo().then(() => {
         INTERVAL_MILLISECONDS
     );
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });