+++ src/jobs/continuous/cancelMoneyTransfer/run.ts
@@ -3,6 +3,7 @@
  */
 import * as pecorino from '@pecorino/domain';
 import * as createDebug from 'debug';
+import * as mongoose from 'mongoose';
 
 import { connectMongo } from '../../../connectMongo';
 
@@ -13,7 +14,7 @@ connectMongo().then(() => {
 
     const MAX_NUBMER_OF_PARALLEL_TASKS = 10;
     const INTERVAL_MILLISECONDS = 500;
-    const taskRepo = new pecorino.repository.Task(pecorino.mongoose.connection);
+    const taskRepo = new pecorino.repository.Task(mongoose.connection);
 
     setInterval(
         async () => {
@@ -27,8 +28,9 @@ connectMongo().then(() => {
                 debug('count:', count);
                 await pecorino.service.task.executeByName(
                     pecorino.factory.taskName.CancelMoneyTransfer
-                )({ taskRepo: taskRepo, connection: pecorino.mongoose.connection });
+                )({ taskRepo: taskRepo, connection: mongoose.connection });
             } catch (error) {
+                // tslint:disable-next-line:no-console
                 console.error(error);
             }
 
@@ -37,6 +39,7 @@ connectMongo().then(() => {
         INTERVAL_MILLISECONDS
     );
 }).catch((err) => {
+    // tslint:disable-next-line:no-console
     console.error('connetMongo:', err);
     process.exit(1);
 });